+++ MathExpression/Source/Parser/MathExpression.swift
@@ -7,7 +7,7 @@ public struct MathExpression {
         case emptyExpression
         case misplacedBrackets
         case unevenOpeningClosingBracketNumber
-        case consecutiveMultiplicativeOperators(String)
+        case invalidConsecutiveOperators(String)
         case startsWithNonSumOrSubtractionOperator(String)
         case endsWithOperator(String)
     }
@@ -51,5 +51,22 @@ public struct MathExpression {
 private extension Array where Element == MathExpression {
     func evaluate() -> [Double] {
         map { $0.evaluate() }
+    }
+}
+
+extension MathExpression.ValidationError: Equatable {
+    public static func == (lhs: MathExpression.ValidationError, rhs: MathExpression.ValidationError) -> Bool {
+        switch (lhs, rhs) {
+        case (.emptyExpression, .emptyExpression),
+            (.misplacedBrackets, .misplacedBrackets),
+            (.unevenOpeningClosingBracketNumber, .unevenOpeningClosingBracketNumber):
+            return true
+        case (.invalidConsecutiveOperators(let lhsValue), .invalidConsecutiveOperators(let rhsValue)),
+            (.startsWithNonSumOrSubtractionOperator(let lhsValue), .startsWithNonSumOrSubtractionOperator(let rhsValue)),
+            (.endsWithOperator(let lhsValue), .endsWithOperator(let rhsValue)):
+            return lhsValue == rhsValue
+        default:
+            return false
+        }
     }
 }