+++ Package.swift
@@ -1,8 +1,11 @@
-// swift-tools-version:4.2
+// swift-tools-version:5.0
 import PackageDescription
 
 let package = Package(
     name: "MathExpression",
+    platforms: [
+        .macOS(.v10_14), .iOS(.v12), .tvOS(.v12)
+    ],
     products: [
         .library(name: "MathExpression", targets: ["MathExpression"]),
     ],