+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+### 1.3.0
+
+* Bumped minimum Xcode version to 13.0+.
+* Bumped Swift version to 5.0 in `Package.swift`.
+* Bumped target platform minimum version to 12.0+ (iOS and tvOS) and 10.14 (macOS).
+* Improved errors thrown when validation fails due to invalid consecutive operators (on some cases the error was thrown but was of different type).
+* Improved tests implementation to check for correct error being thrown.
+* Removed Travis CI integration due to Open Source projects support being dropped.
+* Reimplemented example app using SwiftUI.
+* Improved documentation and added step-by-step installation guide.
+
+#### Breaking changes
+
+* `MathExpression.ValidationError.consecutiveMultiplicativeOperators(String)` has been renamed to `MathExpression.ValidationError.invalidConsecutiveOperators(String)`.
+
 ### 1.2.0
 
 * Added support for [Swift Package Manager](https://swift.org/package-manager/).